move authFetch logic to authContext
