+++ global/hooks/useApplicationsAPI.tsx
@@ -1,7 +1,6 @@
 import { useEffect, useState } from 'react';
 import { AxiosError, AxiosResponse } from 'axios';
 import { APPLICATIONS_PATH } from 'global/constants/internalPaths';
-import fetchAPI from 'global/utils/fetchAPI';
 import { ApplicationsRequestData } from '../../components/pages/Applications/types';
 import useAuthContext from './useAuthContext';
 
@@ -16,28 +15,28 @@ ApplicationsRequestData) => {
   const [response, setResponse] = useState<AxiosResponse | undefined>(undefined);
   const [isLoading, setIsLoading] = useState<boolean>(true);
   const [error, setError] = useState<AxiosError | undefined>(undefined);
-  const { token } = useAuthContext();
+
+  const { fetchWithAuth } = useAuthContext();
 
   useEffect(() => {
-    fetchAPI({
+    fetchWithAuth({
       params: {
         page,
         pageSize,
         sort,
       },
       url: APPLICATIONS_PATH,
-      token,
     })
       .then((res: any) => {
         setResponse(res);
       })
-      .catch((err) => {
+      .catch((err: any) => {
         setError(err);
       })
       .finally(() => {
         setIsLoading(false);
       });
-  }, [page, pageSize, sort, token]);
+  }, [page, pageSize, sort]);
 
   return { error, isLoading, response };
 };
