+++ global/utils/index.ts
@@ -1,1 +1,1 @@
-export { default as fetchAPI } from './fetchAPI';
+export default {};
